LayoutTests/media/track/track-webvtt-tc013-settings-expected.txt

Tests WebVTT settings.


*** Testing text track 0
EXPECTED (cues.length == '4') OK
EXPECTED (cues[0].line == '100') OK
EXPECTED (cues[0].position == '50') OK
EXPECTED (cues[0].align == 'start') OK
EXPECTED (cues[0].vertical == '') OK
EXPECTED (cues[1].line == '15') OK
EXPECTED (cues[1].position == '40') OK
EXPECTED (cues[1].align == 'middle') OK
EXPECTED (cues[1].vertical == 'rl') OK
EXPECTED (cues[2].line == '-1') OK
EXPECTED (cues[2].position == '10') OK
EXPECTED (cues[2].align == 'middle') OK
EXPECTED (cues[2].vertical == '') OK
EXPECTED (cues[3].line == '95') OK
EXPECTED (cues[3].position == '50') OK
EXPECTED (cues[3].align == 'end') OK
EXPECTED (cues[3].vertical == 'lr') OK

*** Testing text track 1
EXPECTED (cues.length == '4') OK
EXPECTED (cues[0].line == '43') OK
EXPECTED (cues[0].position == '10') OK
EXPECTED (cues[0].align == 'middle') OK
EXPECTED (cues[0].vertical == '') OK
EXPECTED (cues[1].line == '-1') OK
EXPECTED (cues[1].position == '50') OK
EXPECTED (cues[1].align == 'end') OK
EXPECTED (cues[1].vertical == '') OK
EXPECTED (cues[2].line == '-1') OK
EXPECTED (cues[2].position == '50') OK
EXPECTED (cues[2].align == 'middle') OK
EXPECTED (cues[2].vertical == '') OK
EXPECTED (cues[3].line == '-1') OK
EXPECTED (cues[3].position == '90') OK
EXPECTED (cues[3].align == 'middle') OK
EXPECTED (cues[3].vertical == 'lr') OK
END OF TEST
